moved svgIcons from utils to components and renamed into SvgIcon
