+++ src/js/components/SurveyCollection.js
@@ -3,7 +3,7 @@ import React, {Fragment} from "react";
 import {removeEnumerator} from "../utils/surveyUtils";
 import {UnicodeIcon} from "../utils/generalUtils";
 import {CollapsibleTitle} from "./FormComponents";
-import {SvgIcon} from "../utils/svgIcons";
+import SvgIcon from "../components/SvgIcon";
 import {mercator} from "../utils/mercator";
 
 export class SurveyCollection extends React.Component {
